# phase1_openapi_v1.yaml
openapi: 3.0.3
info:
  title: Hybrid Fleet Platform API - Phase 1
  version: 1.0.0
  description: "Phase 1 OpenAPI draft aligned to SA / PRD / SD. Scope: Fleet Management Core + Dispatch Compliance Core."
servers:
  - url: https://api.example.com
    description: Production
  - url: https://sandbox-api.example.com
    description: Sandbox
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    TenantApiKey:
      type: apiKey
      in: header
      name: X-API-Key
    WebhookSignature:
      type: apiKey
      in: header
      name: X-Signature
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        requestId:
          type: string
...
    Vehicle:
      type: object
      properties:
        vehicleId:
          type: string
          format: uuid
        plateNo:
          type: string
        vin:
          type: string
        vehicleForm:
          type: string
          enum:
            - sedan
            - wagon
            - mpv
            - accessible
            - other
        licenseClass:
          type: string
          enum:
            - taxi
            - rental
            - multi_taxi
            - other
        energyType:
          type: string
          enum:
            - fuel
            - ev
            - hybrid
            - other
        operatingAreaId:
          type: string
          format: uuid
        dispatchableFlag:
          type: boolean
        currentStatus:
          type: string
        latestReviewStatus:
          type: string
    VehicleCreateRequest:
      type: object
      required:
        - plateNo
        - vin
        - vehicleForm
        - licenseClass
        - operatingAreaId
      properties:
        plateNo:
          type: string
        vin:
          type: string
        vehicleForm:
          type: string
          enum:
            - sedan
            - wagon
            - mpv
            - accessible
            - other
        licenseClass:
          type: string
          enum:
            - taxi
            - rental
            - multi_taxi
            - other
        energyType:
          type: string
          enum:
            - fuel
            - ev
            - hybrid
            - other
        operatingAreaId:
          type: string
          format: uuid
        ownerPartnerId:
          type: string
          format: uuid
        dispatchPartnerId:
          type: string
          format: uuid
        serviceTags:
          type: array
          items:
            type: string
    Driver:
      type: object
      properties:
        driverId:
          type: string
          format: uuid
        name:
          type: string
        mobile:
          type: string
        status:
          type: string
        boundVehicleId:
          type: string
          format: uuid
          nullable: true
        serviceRating:
          type: number
        complaintCount:
          type: integer
    DriverCreateRequest:
      type: object
      required:
        - name
        - mobile
      properties:
        name:
          type: string
        mobile:
          type: string
        occupationalLicenseNo:
          type: string
        occupationalLicenseExpiry:
          type: string
          format: date
        taxiRegistrationNo:
          type: string
        taxiRegistrationExpiry:
          type: string
          format: date
        boundVehicleId:
          type: string
          format: uuid
    VehicleContract:
      type: object
      properties:
        contractId:
          type: string
          format: uuid
        vehicleId:
          type: string
          format: uuid
        partnerId:
          type: string
          format: uuid
        partnerType:
          type: string
        contractType:
          type: string
        serviceScope:
          type: string
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        status:
          type: string
...
paths:
...
  /api/reg/vehicles:
    post:
      tags:
        - RegulatoryVehicles
      summary: Create vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleCreateRequest"
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      tags:
        - RegulatoryVehicles
      summary: List vehicles
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Vehicle"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/reg/drivers:
    post:
      tags:
        - RegulatoryDrivers
      summary: Create driver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverCreateRequest"
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      tags:
        - RegulatoryDrivers
      summary: List drivers
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Driver"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/reg/contracts:
    post:
      tags:
        - Contracts
      summary: Create vehicle contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - vehicleId
                - partnerId
                - partnerType
                - contractType
                - serviceScope
                - startAt
                - endAt
              properties:
                vehicleId:
                  type: string
                  format: uuid
                partnerId:
                  type: string
                  format: uuid
                partnerType:
                  type: string
                contractType:
                  type: string
                serviceScope:
                  type: string
                startAt:
                  type: string
                  format: date-time
                endAt:
                  type: string
                  format: date-time
                operatingAreaId:
                  type: string
                  format: uuid
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleContract"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      tags:
        - Contracts
      summary: List contracts
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/VehicleContract"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...